create-user-key-store.sh: Add gpg 1.x support
Issue: LIN1019-2938

A prior commit that added the interactive interface to the script
removed the gpg 1.x support.  The gpg 1.x has a slightly different
command line syntax.  This commit allows gpg 1.x to work with the
script.

Also added was the GPG_BIN variable.  This allows someone to specify
an alternate GPG binary in the case a system as /usr/bin/gpg and
/usr/bin/gpg2.

The allow-loopback-pinentry needs to be added to the gpg-agent.conf in
order for some versions of gpg 2 to work properly with the passphrase.

(LOCAL REV: NOT UPSTREAM) -- Sent to meta-secure-core on 20190814

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>