tpm2-tools: fix CVE-2021-3565
Issue: LIN1021-355

CVE-2021-3565:
A flaw was found in tpm2-tools in versions before 5.1.1 and before
4.3.2. tpm2_import used a fixed AES key for the inner wrapper,
potentially allowing a MITM attacker to unwrap the inner portion and
reveal the key being imported. The highest threat from this
vulnerability is to data confidentiality.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2021-3565

Patch from:
https://github.com/tpm2-software/tpm2-tools/commit/c069e4f179d5e6653a84fb236816c375dca82515

(LOCAL REV: NOT UPSTREAM) -- Sent to meta-secure-core on 20210707

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>